Backend/Analytics/analytics/requirements.txt
djangorestframework
djongo
pymongo